change disconnection flow

+++ includes/API/Site/Controllers/SnapchatBusinessExtensionController.php
@@ -277,7 +277,6 @@ class SnapchatBusinessExtensionController extends RESTBaseController {
 	 */
 	public function delete_connection() {
 		$config_id = Options::get( OptionDefaults::CONFIG_ID );
-		$config_id = 'f752ecc2-9775-44ee-a5c2-dc3db1b1cfa5';
 
 		if ( $config_id ) {
 			$response = $this->wcs->proxy_delete(
@@ -307,41 +306,51 @@ class SnapchatBusinessExtensionController extends RESTBaseController {
 					500
 				);
 			}
+
+			if ( ! empty( $data['request_status'] ) && 'SUCCESS' === $data['request_status'] ) {
+				$response = $this->stop_connection();
+
+				if ( is_wp_error( $response ) ) {
+					return new WP_REST_Response(
+						array(
+							'status'  => 'error',
+							'message' => $response->get_error_message(),
+							'data'    => $response->get_error_data(),
+						),
+						500
+					);
+				}
+
+				$data         = $response->get_data();
+				$oauth_status = '';
+
+				if ( $data['status'] && 'disconnected' === $data['status'] ) {
+					$oauth_status = $data['status'];
+				}
+
+				Options::delete( OptionDefaults::CONFIG_ID );
+				Options::delete( OptionDefaults::ORGANIZATION_ID );
+				Options::delete( OptionDefaults::ORGANIZATION_NAME );
+				Options::delete( OptionDefaults::AD_ACCOUNT_ID );
+				Options::delete( OptionDefaults::AD_ACCOUNT_NAME );
+				Options::delete( OptionDefaults::CONVERSION_ACCESS_TOKEN );
+				Options::delete( OptionDefaults::PIXEL_ID );
+				Transients::delete( TransientDefaults::PIXEL_SCRIPT );
+
+				return rest_ensure_response(
+					array(
+						'status' => $oauth_status,
+					)
+				);
+			}
 		}
 
-		$response = $this->stop_connection();
-
-		if ( is_wp_error( $response ) ) {
-			return new WP_REST_Response(
-				array(
-					'status'  => 'error',
-					'message' => $response->get_error_message(),
-					'data'    => $response->get_error_data(),
-				),
-				500
-			);
-		}
-
-		$data         = $response->get_data();
-		$oauth_status = '';
-
-		if ( $data['status'] && 'disconnected' === $data['status'] ) {
-			$oauth_status = $data['status'];
-		}
-
-		Options::delete( OptionDefaults::CONFIG_ID );
-		Options::delete( OptionDefaults::ORGANIZATION_ID );
-		Options::delete( OptionDefaults::ORGANIZATION_NAME );
-		Options::delete( OptionDefaults::AD_ACCOUNT_ID );
-		Options::delete( OptionDefaults::AD_ACCOUNT_NAME );
-		Options::delete( OptionDefaults::CONVERSION_ACCESS_TOKEN );
-		Options::delete( OptionDefaults::PIXEL_ID );
-		Transients::delete( TransientDefaults::PIXEL_SCRIPT );
-
-		return rest_ensure_response(
+		return new WP_REST_Response(
 			array(
-				'status' => $oauth_status,
-			)
+				'status'  => 'error',
+				'message' => 'Config id missing',
+			),
+			500
 		);
 	}
